fix/rename Exception To exception
