+++ src/main/java/com/example/projectlxp/service/course/CourseService.java
@@ -1,7 +1,7 @@
 package com.example.projectlxp.service.course;
 
-import com.example.projectlxp.controller.BusinessException;
-import com.example.projectlxp.controller.ExceptionCode;
+import com.example.projectlxp.exception.BusinessException;
+import com.example.projectlxp.exception.ExceptionCode;
 import com.example.projectlxp.model.course.Course;
 import com.example.projectlxp.model.course.CourseStatus;
 import com.example.projectlxp.repository.course.CourseRepository;
